+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - ImageMagick codec can now read and write `gif` and `apng` including animations
 - GIF decoding now uses `-coalesce` to handle partial frames correctly
 - Redimension filter to change image dimensions
+- Redimension attempts to resize the existing buffer when increasing height or
+  adding a new dimension
 
 ## [0.2.0] - 2025-07-21
 - Ruby Sixel encoder now sets pixel aspect ratio metadata to display correctly in Windows Terminal